Add files via upload
tar.gz archives containing the version of the PopART individual-based model (PopART IBM) used to generate transmission chains for the PANGEA phylogenetic methods comparison exercise 2016. The code in this branch is an early version of the PopART IBM, and this code should therefore only be used as part of the PANGEA HIV sequence simulator available from https://github.com/olli0601/PANGEA.HIV.sim
More up-to-date versions of the PopART IBM are available in this repository for general simulation development.